Update core: - Add MetamaskService verify_signature, confirm_user_wallet
Refactor metamask_confirm:
- Add signature verification

Refactor profile.html script;

Update base requirements;

+++ prm/tokens/models.py
@@ -86,7 +86,7 @@ class TokenRound(models.Model):
         return round(self.unit_price * self.total_amount)
 
     @property
-    def progress(self) -> float | int:
+    def progress(self) -> int | float:
         """Подсчитать прогресс текущего раунда в процентах"""
         if self.total_amount:
             return round((self.total_amount_sold / self.total_amount) * 100, 2)
